Read 3D insect brain meshes from insectbraindb.org

Read templatebrains, comprised of their different neuropils, for various insect species from from insectbraindb.org (2018 version), given a single latin names for the species desired. Metadata for available neurons can be seen, and IDs chosen, by calling insectbraindb_species_info. 3D triangular brain meshes are returned as a nat package hxsurf object, which mimics the Amira surface format. These can be be plotted in 3D using rgl and analysed with tools from the nat ecosystem. This incldue subseting by neuropil, i.e.. if you only want to visualise or analyse the antennal lobe.

insectbraindb_read_brain(species = insectbraindb_species_info()$scientific_name,
  brain.sex = c("UNKNOWN", "MALE", "FEMALE"), progress = TRUE)

Arguments

species
the full scientific name for a species. The available options can be seen here

brain.sex
the sex of the species' brain. The available options can be seen here

progress
if TRUE or a numeric value, a progress bar is shown to track the state of your download

Value

a nat package hxsurf object, which mimics the Amira surface format, replete with metadata that can be accessed using $

Details

A single 3D brain object is read, a .obj file for each of its neuropils is downloaded from https://ibdb-file-storage.s3.amazonaws.com/ to a temporary directory, and read using read.obj into a hxsurf object in R, which mimics the Amira surface format. As of May 2019, data from the following species is hosted on insectbraindb.org :

Agrotis infusa Bogong moth

Agrotis segetum Turnip moth

Apis mellifera Honeybee

Apis mellifera Honeybee

Danaus plexippus Monarch Butterfly

Helicoverpa armigera Cotton Bollworm,

Helicoverpa assulta Oriental tobacco budworm

Heliothis virescens Tobacco budworm

Macroglossum stellatarum Hummingbird hawk moth

Manduca sexta Tobacco hornworm

Megalopta genalis Sweat bee

Nasonia vitripennis Jewel wasp

Scarabaeus lamarcki Diurnal dung beetle

Schistocerca gregaria Desert Locust

Note that since neurons are reconstructed from many different neural species, there is no 'standard' orientation between species, but within a species these neurons are registered to a template brain, usually using elastix.

See also

Examples

# NOT RUN {
## What neurons does the insectbraindb.org host?
available.neurons = insectbraindb_neuron_info()

## Let's just download all of the neurons in the database to play with,
## there are not very many:
nrow(available.neurons)

## First, we call the read neurons function, with ids set to NULL
insect.neurons = insectbraindb_read_neurons(ids = NULL)

## Hmm, let's see how many neurons we have perspecies
table(insect.neurons[,"common_name"])

## So, it seem the Monarch Butterfly is the clear winner there, 
## maybe let's just have those
butterfly.neurons =
subset(insect.neurons, common_name == "Monarch Butterfly")

## And let's plot them
plot3d(butterfly.neurons, lwd = 2, soma = 5)

## Cool! But maybe we also want to see it's template brain? 
## Let's check if they have it
available.brains = insectbraindb_species_info()
available.brains

## Great, they do, let's get it
butterfly.brain = insectbraindb_read_brain(species = "Danaus plexippus")

## And plot in a translucent manner
plot3d(butterfly.brain, alpha = 0.1)

## Oop, that's a lot of neuropils. 
## Let's go for only a subset. What's available?
butterfly.brain$RegionList
butterfly.brain$neuropil_full_names

# There lateral horn (LH) and the antennal lobe (AL) are my favourites.
# Let's plot those
clear3d()
plot3d(subset(butterfly.brain, "LH|AL"), alpha = 0.5)

# }