Read insect neurons from insectbraindb.org

Read templatebrain-registered neurons from insectbraindb.org (2018 version), given a single ID or vector of neuron IDs. Alternatively, if ids is set to NULL, all neurons in the database that can be read, are read (~100). Metadata for available neurons can be seen, and IDs chosen, by calling insectbraindb_neuron_info. Neurons are returned as a nat package neuronlist object, that can be be plotted in 3D using rgl and analysed with tools from the nat ecosystem. Each neuron in the insectbraindb is represented by a unique numerical ID, a name and is associated with a species of insect and a publiction, by DOI. Some neurons cannot be read because a SWC file is not available from insectbraindb.org. You can examine the available neurons here here), while the available species may be examined here). Critically, these neurons are registered to standard templates that may be obtained as hxsurf objects in R by using insectbraindb_read_brain.

insectbraindb_read_neurons(ids = NULL, progress = TRUE)

Arguments

ids
a neuron ID, or vector of neuron IDs, as recorded in the insectbraindb.org. If set to NULL (default) then all neurons in the databse are read. As of May 2018, this is only 100 readable neurons. A helpful information page on each neuron can be seen by visiting "https://insectbraindb.org/app/neuron/id", where id is a number, e.g. for Bogon moth neuron 219, this is yuor ticket.

progress
if TRUE or a numeric value, a progress bar is shown to track the state of your download

Value

a nat package neuronlist object, replete with metadata, is returned

Details

Multiple neurons are read, their SWC files as hosted at https://ibdb-file-storage.s3.amazonaws.com/ are downloaded to a temporary directory, and read using read.neurons into a neuronlist object in R. This format and its manipulation is described in detail here. When using insectbraindb_read_neurons, meta data for the neuron is also returned that gives a neuron's ID, short name, long name, associated laboratory and publication, if available, and the assoeciated insect species. As of May 2019, data from the following species is hosted on insectbraindb.org :

Agrotis infusa Bogong moth

Agrotis segetum Turnip moth

Apis mellifera Honeybee

Apis mellifera Honeybee

Danaus plexippus Monarch Butterfly

Helicoverpa armigera Cotton Bollworm,

Helicoverpa assulta Oriental tobacco budworm

Heliothis virescens Tobacco budworm

Macroglossum stellatarum Hummingbird hawk moth

Manduca sexta Tobacco hornworm

Megalopta genalis Sweat bee

Nasonia vitripennis Jewel wasp

Scarabaeus lamarcki Diurnal dung beetle

Schistocerca gregaria Desert Locust

Note that since neurons are reconstructed from many different neural species, there is no 'standard' orientation between species, but within a species these neurons are registered to a template brain, usually using elastix. To visualise the templatebrain, use plot3d on class hxsurf objects downloaded using insectbraindb_read_brain.

See also

Examples

# NOT RUN {
## What neurons does the insectbraindb.org host?
available.neurons = insectbraindb_neuron_info()

## Let's just download all of the neurons in the database to play with,
## there are not very many:
nrow(available.neurons)

## First, we call the read neurons function, with ids set to NULL
insect.neurons = insectbraindb_read_neurons(ids = NULL)

## Hmm, let's see how many neurons we have perspecies
table(insect.neurons[,"common_name"])

## So, it seem the Monarch Butterfly is the clear winner there, 
## maybe let's just have those
butterfly.neurons =
subset(insect.neurons, common_name == "Monarch Butterfly")

## And let's plot them
plot3d(butterfly.neurons, lwd = 2, soma = 5)

## Cool! But maybe we also want to see it's template brain? 
## Let's check if they have it
available.brains = insectbraindb_species_info()
available.brains

## Great, they do, let's get it
butterfly.brain = insectbraindb_read_brain(species = "Danaus plexippus")

## And plot in a translucent manner
plot3d(butterfly.brain, alpha = 0.1)

## Oop, that's a lot of neuropils. 
## Let's go for only a subset. What's available?
butterfly.brain$RegionList
butterfly.brain$neuropil_full_names

# There lateral horn (LH) and the antennal lobe (AL) are my favourites.
# Let's plot those
clear3d()
plot3d(subset(butterfly.brain, "LH|AL"), alpha = 0.5)

# }