Get information on an article by querying an article ID

Returns a data frame describing an article; article_ids can be obtained using neuromorpho_articles_from_neurons

neuromorpho_get_article(
  article_id,
  batch.size = 10,
  neuromorpho_url = "http://neuromorpho.org",
  progress = TRUE,
  ...
)

Arguments

article_id

a unique neuromorpho article ID, number in the repository start from 1. article_ids can be obtained using neuromorpho_articles_from_neurons

batch.size

the number of requests sent at once to the neuromorpho.org, using multi_run. Requests are sent to neuromorpho.org in parallel to speed up the process of reading neurons. Batches of queries are processed serially. Increasing the value of batch.size may reduce read time.

neuromorpho_url

the base URL for querying the neuromorpho database, defaults to http://neuromorpho.org

progress

if TRUE or a numeric value, a progress bar is shown. The bar progresses when each batch is completed. If TRUE, or 100, the bar completes where all batches are done.

...

methods passed to neuromorpho_async_req, or in some cases, neuromorphr:::neuromorpho_fetch

Details

Articles can also be searched in a web browser at http://neuromorpho.org/LS.jsp.

See also