Convert neuromorpho neuron IDs into neuron names, or other information, and vica versa

Every neuron in the neuromorpho repository has its own unique neuron ID, starting at 1, and name. This can be mapped to a neuron name (also unique), or another (non-unique) neuron field entry. Names and IDs can be used to read neurons from neuromorpho.org, via neuromorpho_read_neurons.

neuromorpho_names_from_ids(
  neuron_id,
  progress = TRUE,
  batch.size = 10,
  neuromorpho_url = "http://neuromorpho.org",
  ...
)

neuromorpho_ids_from_names(
  neuron_name,
  progress = TRUE,
  batch.size = 10,
  neuromorpho_url = "http://neuromorpho.org",
  ...
)

neuromorpho_field_entries_from_names(
  neuron_name,
  progress = TRUE,
  batch.size = 10,
  neuromorpho_url = "http://neuromorpho.org",
  field = "archive",
  ...
)

neuromorpho_articles_from_neurons(
  neuron_name = NULL,
  neuron_id = NULL,
  progress = TRUE,
  search.type = c("doi", "pmid"),
  neuromorpho_url = "http://neuromorpho.org",
  ...
)

Arguments

neuron_id

a neuron ID, or vector of neuron IDs, as recorded in the neuromorpho database. If neuron_name is given this supersedes neuron_id, which is then treated as if its value were NULL.

progress

if TRUE or a numeric value, a progress bar is shown. The bar progresses when each batch is completed. If TRUE, or 100, the bar completes where all batches are done.

batch.size

the number of requests sent at once to the neuromorpho.org, using multi_run. Requests are sent to neuromorpho.org in parallel to speed up the process of reading neurons. Batches of queries are processed serially. Increasing the value of batch.size may reduce read time.

neuromorpho_url

the base URL for querying the neuromorpho database, defaults to http://neuromorpho.org

...

methods passed to neuromorpho_async_req, or in some cases, neuromorphr:::neuromorpho_fetch

neuron_name

a neuron name, or vector of neuron names, as recorded in the neuromorpho database. Names and neuron IDs can be found by searching the repository, for example via neuromorpho_search

field

another field for neurons' metadata, i.e. one of the fields as returned by neuromorpho_fields

search.type

search for a publication associated with a neuron, either by looking for a corresponding doi or pmid (PubMed ID).

Value

a vector of field entries corresponding to the given neuron IDs / names

Details

retrieves the meta data associated with a neuron ID / neuron name, in order to find its corresponding name / ID. This is useful because the name can be ued to directly locate a neuron's SWC file, but the neuron ID cannot.

See also

Examples

if (FALSE) {
# find the first 10 neurons uploaded to the neuromorpho respository
neuron_names = neuromorpho_names_from_ids(c(1:10))

# and see that we get the same IDs back if we search the other way
neuron_ids = neuromorpho_ids_from_names(neuron_names)

# now we can see the corresponding article IDs
article_ids = neuromorpho_articles_from_neurons(neuron_names, search.type = "doi")

}