Get all meta data associated with a neuron

Retrieve the meta data associated with query neurons. Neurons can be queried using either their neuron ID or their name.

neuromorpho_neurons_info(
  neuron_name = NULL,
  neuron_id = NULL,
  batch.size = 10,
  progress = TRUE,
  neuromorpho_url = "http://neuromorpho.org",
  ...
)

Arguments

neuron_name

a neuron name, or vector of neuron names, as recorded in the neuromorpho database. Names and neuron IDs can be found by searching the repository, for example via neuromorpho_search

neuron_id

a neuron ID, or vector of neuron IDs, as recorded in the neuromorpho database. If neuron_name is given this supersedes neuron_id, which is then treated as if its value were NULL.

batch.size

the number of requests sent at once to the neuromorpho.org, using multi_run. Requests are sent to neuromorpho.org in parallel to speed up the process of reading neurons. Batches of queries are processed serially. Increasing the value of batch.size may reduce read time.

progress

if TRUE or a numeric value, a progress bar is shown. The bar progresses when each batch is completed. If TRUE, or 100, the bar completes where all batches are done.

neuromorpho_url

the base URL for querying the neuromorpho database, defaults to http://neuromorpho.org

...

methods passed to neuromorpho_async_req, or in some cases, neuromorphr:::neuromorpho_fetch

Value

a list containing all the meta data for a neuron stored on neuromorpho.org

Details

The meta data for reconstructions has been extracted by neuromorpho.org from corresponding publications, and generally include:

Web URL of archives (if available) with any additional information about the reconstruction

Subject-related information - species, strain, age, weight, sex

Experiment-related information - protocol, experimental condition, staining method, slicing direction and thickness, objective type and magnification.

Brain region at three levels: main region/structure, sub-region, detailed sub-region (example: Hippocampus, dentate gyrus, granule cell layer)

Neuron type at three levels: main neuron class type, sub-class, detailed sub-class (example: Interneuron, basket cell, nested)

Reconstruction method

Format of original data, though currently this package only retrieves the data as standardised by neuromorpho.org

See also

Examples

if (FALSE) { 
# These are some names for a few elephant neurons
elephant_neuron_names = c("155-2-1Stel", "155-2-2Stel", "155-2-5Stel")

# We can now get the meta information associated with these reconstructions
elephant_neuron_info = neuromorpho_neurons_info(elephant_neuron_names)

# We can get essentially the same information as a data frame instead
elephant_neuron_meta = neuromorpho_neurons_meta(elephant_neuron_names)
}